Fixes release build of server causing stackoverflow on start
